+++ ui/submodules/packages/components/src/scenarios/index.ts
@@ -1,3 +1,2 @@
 export { RootRoutes } from "./rootRoutes";
 export { ScenariosView } from "./scenariosView";
-export { useScenariosWithStatus, useUserQuery } from "./useScenariosQuery";
